+++ CHANGELOG.md
@@ -12,6 +12,11 @@ All notable changes to this project are documented in this file.
 
 - (INTERNAL) Renamed certain functions to better convey their meaning. [#5]
 
+### Fixed
+
+- Bug that resulted in query parameters not being properly handled when wrapped within
+  `Bulk`. [#7]
+
 
 ## 0.1.1 [2025-11-26]
 
@@ -23,3 +28,4 @@ All notable changes to this project are documented in this file.
 
 [#5]: https://github.com/manoss96/onlymaps/pull/5
 [#6]: https://github.com/manoss96/onlymaps/pull/6
+[#7]: https://github.com/manoss96/onlymaps/pull/7